geting one item

+++ src/todo/todo.service.ts
@@ -1,22 +1,23 @@
 import { Injectable } from '@nestjs/common';
-import { TodoItemDto } from 'src/dto/NewTodoItem.dto';
-import { SetItemStatusDto } from 'src/dto/SetItemStatusDto.dto';
+import { TodoItemDto } from '../dto/NewTodoItem.dto';
+import { SetItemStatusDto } from '../dto/SetItemStatusDto.dto';
+import { Filters } from '../interfaces/filters.enum';
 import { ToDoListDB } from './ToDoListDB.class';
 
 let listdb = new ToDoListDB([]);
-enum Filters {
-  finished = 'finished',
-  unfinished = 'unfinished',
-}
 
 @Injectable()
 export class TodoService {
-  getAllItems(filter?: Filters) {
+  getAllItems(filter?: Filters): TodoItemDto | TodoItemDto[] {
     if (filter) {
       return listdb.getAll(filter);
     } else {
       return listdb.getAll();
     }
+  }
+
+  getOneItem(id: string) {
+    return listdb.getItem(id);
   }
 
   setItem(newTodoItemDto: TodoItemDto) {
@@ -25,16 +26,24 @@ export class TodoService {
     } else {
       listdb = listdb.saveItem(newTodoItemDto);
     }
-    return true;
+    return listdb.getItem(newTodoItemDto.id);
   }
 
-  removeItem(id: string) {
-    listdb.removeItem(id);
-    return true;
+  removeItem(id: string): boolean {
+    if (listdb.getItem(id)) {
+      listdb.removeItem(id);
+      return true;
+    } else {
+      return false;
+    }
   }
 
-  setItemStatus(setItemStatusDto: SetItemStatusDto) {
-    listdb = listdb.setItemStatus(setItemStatusDto);
-    return true;
+  setItemStatus(setItemStatusDto: SetItemStatusDto): boolean {
+    if (listdb.getItem(setItemStatusDto.id)) {
+      listdb = listdb.setItemStatus(setItemStatusDto);
+      return true;
+    } else {
+      return false;
+    }
   }
 }
